requirements.txt
playwright>=1.44.0
flask==3.0.2
apscheduler==3.10.4
Pillow>=10.3.0
imagehash==4.3.1
pytz==2024.1
